Fix currencies endpoint 404 and the type-mismatch 500 behind it
CurrencyService called the non-existent /api/currencies — the real
endpoint lives on ConfigurationController at
/api/configuration/currencies. Fixing the URL exposed a second,
unrelated bug: Currency.DecimalPrecision was declared as C# int, but
the live database column is tinyint, so any query projecting that
column (this endpoint, and CurrencyConversionService.GetPrecisionAsync
for currency conversion) threw an InvalidCastException. Changed the
type to byte to match reality and applied the aligning migration.

+++ src/WebProlific.Core/Entities/Currency.cs
@@ -6,6 +6,11 @@ public class Currency
     public string Code { get; set; } = string.Empty; // ISO 4217 code, e.g., USD
     public string Name { get; set; } = string.Empty; // Friendly name
     public string Symbol { get; set; } = string.Empty; // Currency symbol
-    public int DecimalPrecision { get; set; } = 2; // Number of fractional digits
+    // byte, not int — the live DB column is tinyint (a value like 2-4 never needs
+    // more range than that); the EF model previously declared int, mismatching the
+    // physical schema and crashing every read of this column with an
+    // InvalidCastException ("Unable to cast object of type 'System.Byte' to
+    // type 'System.Int32'") the moment it was projected instead of the full entity.
+    public byte DecimalPrecision { get; set; } = 2; // Number of fractional digits
     public bool IsActive { get; set; } = true;
 }
